basic river provider wip

+++ island/src/main/java/ca/mcmaster/cas/se2aa4/a3/island/river/RiverProvider.java
@@ -4,5 +4,5 @@ import ca.mcmaster.cas.se2aa4.a3.island.utils.Segment;
 
 public interface RiverProvider {
     // 0: no river
-    double isRiver(Segment segment);
+    int isRiver(Segment segment);
 }
